requirements.txt
streamlit

sqlalchemy
pymysql
python-dotenv
plotly